requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
behave==1.2.6
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
colorama==0.4.6
cssselect==1.2.0
exceptiongroup==1.2.2
execnet==2.1.1
fasteners==0.19
filelock==3.16.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
mycdp==1.0.1
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.4
pdbp==1.6.0
platformdirs==4.3.6
pluggy==1.5.0
py==1.11.0
pycparser==2.22
Pygments==2.18.0
pynose==1.5.3
pyotp==2.9.0
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.3.3
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.3
sbvirtualdisplay==1.3.0
selenium==4.26.1
seleniumbase==4.32.6
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tabcompleter==1.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
websockets==13.1
wheel==0.44.0
wsproto==1.2.0
